refactor(clipboard): one copy helper, native plugin included
Review on c14fdbb5e: CopyToClipboard bypassed the canonical
copyTextToClipboard in clipboard.utils, which already routes native copies
through @capacitor/clipboard, so the Capacitor failure this PR targets could
still end as a failed copy. Fold the timeout race, execCommand fallback and
Sentry reporting into that helper, delete the weaker duplicate in
general.utils, and point its five callers at the canonical one.
