+++ src/utils/__tests__/clipboard.utils.test.ts
@@ -1,4 +1,9 @@
+import * as Sentry from '@sentry/nextjs'
 import { beginClipboardCopy, copyTextToClipboard } from '../clipboard.utils'
+
+jest.mock('@sentry/nextjs', () => ({ captureException: jest.fn(), captureMessage: jest.fn() }))
+// clipboard.utils reports through the lazy wrapper; aliasing it to the mock keeps the calls synchronous.
+jest.mock('@/utils/sentry-lazy', () => require('@sentry/nextjs'))
 
 const mockIsNativeBridge = jest.fn<boolean, []>()
 jest.mock('@/utils/capacitor', () => ({
@@ -129,9 +134,50 @@ describe('copyTextToClipboard', () => {
         expect(writeText).toHaveBeenCalledWith(LINK)
     })
 
-    it('reports failure when the write is refused', async () => {
+    it('reports failure when the write is refused and the fallback cannot copy', async () => {
         writeText.mockRejectedValue(new Error('denied'))
+        document.execCommand = jest.fn(() => false)
 
         await expect(copyTextToClipboard(LINK)).resolves.toBe(false)
+        expect(Sentry.captureException).toHaveBeenCalledTimes(1)
+        expect(Sentry.captureMessage).toHaveBeenCalledTimes(1)
+    })
+
+    // The Capacitor WebView rejects writeText (NotAllowedError) even when the
+    // legacy command still works.
+    it('falls back to execCommand when the write is refused', async () => {
+        writeText.mockRejectedValue(new Error('denied'))
+        document.execCommand = jest.fn(() => true)
+
+        await expect(copyTextToClipboard(LINK)).resolves.toBe(true)
+        expect(document.execCommand).toHaveBeenCalledWith('copy')
+        expect(Sentry.captureException).toHaveBeenCalledTimes(1)
+    })
+
+    // Brave iOS never settles the writeText promise.
+    it('falls back when the write never settles', async () => {
+        jest.useFakeTimers()
+        try {
+            writeText.mockReturnValue(new Promise(() => {}))
+            document.execCommand = jest.fn(() => true)
+
+            const result = copyTextToClipboard(LINK)
+            await jest.advanceTimersByTimeAsync(1000)
+
+            await expect(result).resolves.toBe(true)
+            expect((Sentry.captureException as jest.Mock).mock.calls[0][0].message).toContain('timed out')
+        } finally {
+            jest.useRealTimers()
+        }
+    })
+
+    it('leaves no textarea behind when execCommand throws', async () => {
+        writeText.mockRejectedValue(new Error('denied'))
+        document.execCommand = jest.fn(() => {
+            throw new Error('execCommand unsupported')
+        })
+
+        await expect(copyTextToClipboard(LINK)).resolves.toBe(false)
+        expect(document.querySelector('textarea')).toBeNull()
     })
 })
